:recycle: :: 타입 변경 Int -> Long

+++ src/main/kotlin/com/project/indistraw/domain/movie/adapter/input/data/request/UpdateMovieRequest.kt
@@ -14,9 +14,9 @@ data class UpdateMovieRequest(
     @field:NotNull
     val movieHighlight: List<String>,
     @field:NotNull
-    val director: List<Int>,
+    val director: List<Long>,
     @field:NotNull
-    val actor: List<Int>,
+    val actor: List<Long>,
     @field:NotNull
     val clowdTrue: Boolean
 )
